fix: Always apply containerd patches
This commit ensures that applying containerd patches always happens. Previously,
this only happened if mirror configuration was provided, which meant that patches
such as enabling containerd metrics were never applied to containerd config.

By bundling the apply patches and restart containerd script in the same handler,
we ensure they both always get run.
