+++ pkg/handlers/generic/mutation/containerdapplypatchesandrestart/apply_patches.go
@@ -1,6 +1,6 @@
 // Copyright 2023 Nutanix. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
-package containerdrestart
+package containerdapplypatchesandrestart
 
 import (
 	_ "embed"
