+++ pkg/handlers/generic/mutation/containerdapplypatchesandrestart/templates/containerd-apply-patches.sh.gotmpl
