+++ pkg/handlers/generic/mutation/containerdapplypatchesandrestart/templates/containerd-restart.sh
