+++ pkg/handlers/generic/mutation/mirrors/inject.go
@@ -101,7 +101,7 @@ func (h *globalMirrorPatchHandler) Mutate(
 			if err != nil {
 				return err
 			}
-			files, commands, generateErr := generateFilesAndCommands(
+			files, generateErr := generateFilesAndCommands(
 				mirrorConfig,
 				globalMirror)
 			if generateErr != nil {
@@ -115,15 +115,6 @@ func (h *globalMirrorPatchHandler) Mutate(
 			obj.Spec.Template.Spec.KubeadmConfigSpec.Files = append(
 				obj.Spec.Template.Spec.KubeadmConfigSpec.Files,
 				files...,
-			)
-
-			log.WithValues(
-				"patchedObjectKind", obj.GetObjectKind().GroupVersionKind().String(),
-				"patchedObjectName", ctrlclient.ObjectKeyFromObject(obj),
-			).Info("adding PreKubeadmCommands to control plane kubeadm config spec")
-			obj.Spec.Template.Spec.KubeadmConfigSpec.PreKubeadmCommands = append(
-				obj.Spec.Template.Spec.KubeadmConfigSpec.PreKubeadmCommands,
-				commands...,
 			)
 
 			return nil
@@ -143,7 +134,7 @@ func (h *globalMirrorPatchHandler) Mutate(
 			if err != nil {
 				return err
 			}
-			files, commands, generateErr := generateFilesAndCommands(
+			files, generateErr := generateFilesAndCommands(
 				mirrorConfig,
 				globalMirror)
 			if generateErr != nil {
@@ -156,12 +147,6 @@ func (h *globalMirrorPatchHandler) Mutate(
 			).Info("adding global registry mirror files to worker node kubeadm config template")
 			obj.Spec.Template.Spec.Files = append(obj.Spec.Template.Spec.Files, files...)
 
-			log.WithValues(
-				"patchedObjectKind", obj.GetObjectKind().GroupVersionKind().String(),
-				"patchedObjectName", ctrlclient.ObjectKeyFromObject(obj),
-			).Info("adding PreKubeadmCommands to worker node kubeadm config template")
-			obj.Spec.Template.Spec.PreKubeadmCommands = append(obj.Spec.Template.Spec.PreKubeadmCommands, commands...)
-
 			return nil
 		}); err != nil {
 		return err
@@ -173,11 +158,11 @@ func (h *globalMirrorPatchHandler) Mutate(
 func generateFilesAndCommands(
 	mirrorConfig *mirrorConfig,
 	globalMirror v1alpha1.GlobalImageRegistryMirror,
-) ([]bootstrapv1.File, []string, error) {
+) ([]bootstrapv1.File, error) {
 	// generate default registry mirror file
 	files, err := generateGlobalRegistryMirrorFile(mirrorConfig)
 	if err != nil {
-		return nil, nil, err
+		return nil, err
 	}
 	// generate CA certificate file for registry mirror
 	mirrorCAFile := generateMirrorCACertFile(mirrorConfig, globalMirror)
@@ -185,14 +170,6 @@ func generateFilesAndCommands(
 	// generate Containerd registry config drop-in file
 	registryConfigDropIn := generateContainerdRegistryConfigDropInFile()
 	files = append(files, registryConfigDropIn...)
-	// generate Containerd apply patches script and command
-	var commands []string
-	applyPatchesFile, applyPatchesCommand, err := generateContainerdApplyPatchesScript()
-	if err != nil {
-		return nil, nil, err
-	}
-	files = append(files, applyPatchesFile...)
-	commands = append(commands, applyPatchesCommand)
 
-	return files, commands, err
+	return files, err
 }
