+++ pkg/handlers/generic/mutation/mirrors/mirror_test.go
@@ -174,41 +174,3 @@ func Test_generateContainerdRegistryConfigDropInFile(t *testing.T) {
 	file := generateContainerdRegistryConfigDropInFile()
 	assert.Equal(t, want, file)
 }
-
-func Test_generateContainerdApplyPatchesScript(t *testing.T) {
-	wantFile := []cabpkv1.File{
-		{
-			Path:        "/etc/containerd/apply-patches.sh",
-			Owner:       "",
-			Permissions: "0700",
-			Encoding:    "",
-			Append:      false,
-			//nolint:lll // just a long string
-			Content: `#!/bin/bash
-set -euo pipefail
-IFS=$'\n\t'
-
-declare -r TOML_MERGE_IMAGE="ghcr.io/mesosphere/toml-merge:v0.2.0"
-
-if ! ctr --namespace k8s.io images check "name==${TOML_MERGE_IMAGE}" | grep "${TOML_MERGE_IMAGE}" >/dev/null; then
-  ctr --namespace k8s.io images pull "${TOML_MERGE_IMAGE}"
-fi
-
-cleanup() {
-  ctr images unmount "${tmp_ctr_mount_dir}" || true
-}
-
-trap 'cleanup' EXIT
-
-readonly tmp_ctr_mount_dir="$(mktemp -d)"
-
-ctr --namespace k8s.io images mount "${TOML_MERGE_IMAGE}" "${tmp_ctr_mount_dir}"
-"${tmp_ctr_mount_dir}/usr/local/bin/toml-merge" -i --patch-file "/etc/containerd/cre.d/*.toml" /etc/containerd/config.toml
-`,
-		},
-	}
-	wantCmd := "/bin/bash /etc/containerd/apply-patches.sh"
-	file, cmd, _ := generateContainerdApplyPatchesScript()
-	assert.Equal(t, wantFile, file)
-	assert.Equal(t, wantCmd, cmd)
-}
